agents/rag/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.83.0
anyio==4.12.1
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
brotli==1.2.0
build==1.4.0
bytecode==0.17.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.5.1
click==8.3.1
cssselect2==0.9.0
ddtrace==4.5.1
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
durationpy==0.10
envier==0.6.1
fastapi==0.133.0
filelock==3.24.3
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
googleapis-common-protos==1.72.0
greenlet==3.3.2
grpcio==1.78.1
h11==0.16.0
hf-xet==1.3.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
jiter==0.13.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
Markdown==3.10.2
markdown-it-py==4.0.0
markdownify==1.2.2
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
numpy==2.4.2
oauthlib==3.3.1
onnxruntime==1.24.2
openai==2.24.0
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
overrides==7.7.0
packaging==26.0
pillow==12.1.1
playwright==1.58.0
pluggy==1.6.0
posthog==5.4.0
protobuf==6.33.5
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydyf==0.12.1
pyee==13.0.1
Pygments==2.19.2
pyphen==0.17.2
PyPika==0.51.1
pyproject_hooks==1.2.0
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.30.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
tiktoken==0.12.0
tinycss2==1.5.1
tinyhtml5==2.0.0
tokenizers==0.22.2
tqdm==4.67.3
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
weasyprint==68.1
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
wrapt==2.1.2
zipp==3.23.0
zopfli==0.4.1
